Update Explainer.md
Remove "(from the WoT Architecture document)" when referring to the figure since this exact figure does not appear in the Architecture document any more.   In a separate edit I'll update image links to point to the TAG-review version of the Arch doc to protect against someone accidentally deleting this image.