Copy "building blocks" image from wot-architecture
I am copying this image over since I'm a little afraid someone will delete it in wot-architecture during housekeeping, as this particular image is no longer being used in the wot-architecture document.  I plan to update the link to the image as well so it refers the local copy.  The content of the explainer will not change, I'm just heading off a potential problem by making the image "local".