Create security_and_privacy.md
Initial checkin.  This is just a copy of what we submitted last time, and still needs to be updated.
Here is the original: https://github.com/w3c/wot-architecture/blob/main/proposals/security_and_privacy.md
Note that it was in architecture but was written to relate to both (we should update to just relate to the TD).